Added some components and designs
Added fetch API in the handleSubmit to the log in form but no backend url.
Improvised the sidebar panel. Using useEffect (swith-case) instead of useState handleClick. Since the button needed to double click for the background color of the button to update.
Edited the dashboard but still not final.
Made the Scan QR Code a toggle button whether scan thru cam or scan thru browse file and drag and drop to restore some space in the webpage.
Added state variable and function in Setup Unit, category and supplier.
Added a button where when clicked, it will add field of row. It is in the second table blue in the setup unit.
Added delete row button if the field of row is not needed.
Added search bar for user in the setup computer set.
Added Extract.jsx component for the Qr code result after scanning. This is not yet done.