Edited Dashboard and Updating QR Code
Installed QR Code reader different package (not Beta) .
Installed React QR SVG for generating QR, did not uninstall the React-Qr-Code package, might use it in the future if better preference.
Edited the Tables in Set and List of Computers, added Table Pagination.
Edited the Sidebar, making it smaller.
Added Functions to the Action buttons.
Added Validation in Sign Up and Log In.
Edited the View and EditView in Computer component.
Added Sample Data if needed.
Added data for Position in Computer component.
Refactor code in Extract.jsx (Extract = Page after scanning the Qr Code where to add remarks/update)
Still editing the QR Code Part
Understanding/Plan: (Upon clicking the QR Code button in the Computer Table, it will pop up the generated code. The data of the generated code is the Computer ID of its row. [Computer ID = Unique] Upon scanning the Generated QR Code, the data scanned will toggle below the scanner. It is the Computer ID and it is clickable. Upon clicking it, it will redirect to a page where the specs and details of the computer were displayed. There will be 2 inputs/dropdown. [Remarks and Installed Applications]. Need to sign in/up to submit input.