Db2, Unit, Editset, Signup, View, Deleted the EditView, Qr
Db2- Edited the row-span of the Pie chart for it to have no gap below.
Unit- Edited the table of the set up unit, the dropdown should be visible and not overflow-scroll. (Not yet- Will edit the radius of the top of the table.)
Editset- Changing the position of the name of the use of the computer set, be it beside the buttons and not above to refrain overflow of white paper/div.
Added option 'Delete' in reason of deletion.
Sign-up- Deleting the excess border-color if error occurred since the border color was called twice in the if statement, the other border color should be in the else.
View- Added dialog, the editview dialog.
Deleted the EditView.jsx
Set- Added Remove option in reason of deletion.
Qr- Change the position of the close icon inside the modal.