allUnits, Unit, Specs, Qr, Set
allUnits- Added Table 	Pagination with Empty Rows.
Unit- Added Table Pagination for the Vacant Status Units Table.
Change the border radius of the top of the 2 tables.
Specs- Change the CloseIcon of the modal.
Qr- Change the CloseIcon of the modal.
Set- Added show more: show less button for the table for the rows not to flood.